requirements.txt
# web server
Flask==0.10.1

# orm
Flask-SQLAlchemy==2.1
Flask-Migrate==1.6.0
psycopg2==2.6.1

# misc
Flask-Login==0.3.2
Flask-Cors==2.1.2
python-dateutil==2.4.2
redis==2.10.5
marshmallow==2.3.0
marshmallow-sqlalchemy==0.6.0
